Added tests
Author: Sanketh Mopuru (sanketh.mopuru@gmail.com)

+++ tests/helpers.bash
@@ -1,7 +1,11 @@
 setUpEnv(){
     export TAGS_HOME=$(mktemp -d)
+    export TMPDIR=$(mktemp -d)
 }
 
 tearDownEnv(){
-   rm -rf $TAGS_HOME
+    if [[ $BATS_TEST_COMPLETED ]]; then
+        rm -rf $TAGS_HOME
+        rm -rf $TMPDIR
+    fi
 }
